+++ MordoEngine/src/Engine.cpp
@@ -1,11 +1,12 @@
 #include "Engine.h"
-#include <random>
 #include "API/OpenGL/OpenGLBackend.h"
 #include "Input/Input.h"
-#include "Renderer/Cube/CubeRenderer.h"
+#include "Scene/Scene.h"
+#include "Scene/GameScene.h"
 #include "Core/Managers/Manager.h"
 #include "Core/Texture/Texture.h"
 #include "Core/Shader/Shader.h"
+#include <memory>
 
 void Engine::Run()
 {
@@ -13,12 +14,13 @@ void Engine::Run()
 	Input::Init(OpenGLBackend::GetGLFWwindow());
 	Manager<Texture>::Init();
 	Manager<Shader>::Init();
-	CubeRenderer cubeRenderer(100);
+	std::unique_ptr<Scene> gameScene = std::make_unique<GameScene>();
 
 	float lastTime = 0.0f;
 
 	while (!OpenGLBackend::WindowShouldClose())
 	{
+		static bool drawModeInLines = false;
 		float currentTime = glfwGetTime();
 		float deltaTime = currentTime - lastTime;
 		lastTime = currentTime;
@@ -30,8 +32,18 @@ void Engine::Run()
 			OpenGLBackend::CloseWindow();
 		}
 
-		cubeRenderer.Update(deltaTime);
-		cubeRenderer.Render();
+		if (Input::KeyPressed(GLFW_KEY_SPACE)) {
+			drawModeInLines = !drawModeInLines;
+			if (drawModeInLines) {
+				glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
+			}
+			else {
+				glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
+			}
+		}
+
+		gameScene->Update(deltaTime);
+		gameScene->Render();
 
 		OpenGLBackend::SwapBuffers();
 	}