+++ MordoEngine/src/Renderer/Renderer.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class Renderer
+{
+public:
+	virtual void Render() = 0;
+	virtual ~Renderer() = default;
+};