+++ MordoEngine/src/Terrain/HeightMapTerrain.cpp
@@ -1,0 +1,6 @@
+#include "HeightMapTerrain.h"
+
+HeightMapTerrain::HeightMapTerrain(const std::string& filepath)
+{
+	LoadHeightMap(filepath);
+}