+++ MordoEngine/src/Terrain/HeightmapTerrain.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "Terrain.h"
+#include <string>
+
+class HeightMapTerrain : public terrain::Terrain 
+{
+public:
+	HeightMapTerrain(const std::string& filepath);
+};